requirements-dev.txt
-r requirements.txt
pytest>=8.0
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
ruff>=0.5
